+++ src/_types/header_row_T.d.ts
@@ -1,8 +1,12 @@
+import type { row_T } from './row_T'
 export type header_row_T<
 	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
 > =
-	ColDefs extends [string, any][]
-	? { [K in keyof ColDefs]:ColDefs[K][0] }
-	: ColDefs extends any[]
-		? ColDefs
-		: string[]
+	row_T<ColDefs>
+	&(ColDefs extends [string, any][]
+		? { [K in keyof ColDefs]:ColDefs[K][0] }
+		: ColDefs extends any[]
+			? ColDefs
+			: ColDefs extends object
+				? KeyOf<ColDefs>[]
+				: string[])