+++ src/_types/row_T.d.ts
@@ -1,0 +1,10 @@
+export type row_T<
+	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
+> =
+	ColDefs extends [string, any][]
+	? { [K in keyof ColDefs]:any }
+	: ColDefs extends any[]
+		? ColDefs
+		: ColDefs extends object
+			? any[]
+			: string[]